observability-stack/payment-api-mock/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
influxdb-client==1.39.0
prometheus-fastapi-instrumentator==6.2.0
python-dotenv==1.0.1
pydantic==2.6.4
